refactor: 리펙토링 - 프리티어 재설정 - style.ts -> styled.tsx로 변경
