+++ client/src/components/modules/ProductCard/ProductCard.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import * as S from './style';
+import * as S from './styled';
 import { IconType } from '@utils/constants';
 import Icon from '@components/atoms/Icon';
 
@@ -18,11 +18,17 @@ export const ProductCard: React.FC<Props> = ({ name, price }) => {
   return (
     <S.ProductCard>
       <S.ProductImg width={200} height={200} src={require('../../../assets/images/coke.jpeg')}>
-        <div onClick={onLikeHandler}>{Liked ? <Icon icon={IconType.HEART} size={1.5} /> : <Icon icon={IconType.REG_HEART} size={1.5} />}</div>
+        <div onClick={onLikeHandler}>
+          {Liked ? (
+            <Icon icon={IconType.HEART} size={1.5} />
+          ) : (
+            <Icon icon={IconType.REG_HEART} size={1.5} />
+          )}
+        </div>
       </S.ProductImg>
       <div>
-        <div className='item-name'>{name}</div>
-        <div className='item-price'>{price}원</div>
+        <div className="item-name">{name}</div>
+        <div className="item-price">{price}원</div>
       </div>
     </S.ProductCard>
   );
